# src/main/java/com/shashtra/openapidemo/controller/data/GetUserResponse.java
package com.shashtra.openapidemo.controller.data;

import com.shashtra.openapidemo.controller.request.Userdata;

/***************
 **  author: Hariom
 **  since: 23/07/22
 *****************/
public class GetUserResponse extends Userdata {
  private String userId;
}