# src/main/java/com/shashtra/openapidemo/controller/request/Userdata.java
package com.shashtra.openapidemo.controller.request;

/***************
 **  author: Hariom
 **  since: 23/07/22
 *****************/
public class Userdata {
    private String firstName;
    private String lastName;

    private String emailAddress;
}