# src/main/java/com/shashtra/openapidemo/error/ErrorResponse.java
package com.shashtra.openapidemo.error;

/***************
 **  author: Hariom
 **  since: 23/07/22
 *****************/
public class ErrorResponse {
    private String code;
    private String message;
}